Go: Improve Parallel Letter Frequency
It seems recursively dividing-and-conquering is intended to be the solution for this, although I'm not really sure why it's a better approach than sequentially kicking off the goroutines?

How long does a list need to get that splitting it in half is quicker than sequentially looping through?

That's a question I could try and answer with code...